requirements.txt
discord.py
git+https://github.com/nafiz1001/cowsay-python
